request controller update according branch manager

+++ app/controllers/requests_controller.rb
@@ -10,6 +10,8 @@ class RequestsController < ApplicationController
     @sub_counties = SubCounty.none
     @requests = Request.apply_filters(params).order(created_at: :desc)
     @requests = @requests.where(event_id: nil) if params.except(:controller, :action).empty?
+     # Restrict to branch_manager's branch
+    @requests = @requests.where(branch_id: current_user.branch_id) if current_user.role == 'branch_manager'
   end
 
   def ussd
@@ -45,6 +47,8 @@ class RequestsController < ApplicationController
 
   def create
     @request = Request.new(request_params)
+     # Restrict branch_id for branch_manager
+    @request.branch_id = current_user.branch_id if current_user.role == 'branch_manager'
 
     if @request.save
       redirect_to @request, notice: 'Request was successfully created.'
@@ -56,6 +60,10 @@ class RequestsController < ApplicationController
   end
 
   def update
+    # Prevent branch_manager from updating requests outside their branch
+    unless @request.branch_id == current_user.branch_id || current_user.admin? || current_user.super_admin?
+      redirect_to requests_path, alert: 'You are not authorized to update this request.' and return
+    end
     if @request.update(request_params)
       redirect_to @request, notice: 'Request was successfully updated.'
     else
@@ -68,6 +76,10 @@ class RequestsController < ApplicationController
   end
 
   def destroy
+    # Prevent branch_manager from deleting requests
+    unless @request.branch_id == current_user.branch_id || current_user.admin? || current_user.super_admin?
+      redirect_to requests_path, alert: 'You are not authorized to delete this request.' and return
+    end
     if @request.destroy
       redirect_to requests_url, notice: 'Request was successfully destroyed.'
     else
